backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
openai==1.12.0
anthropic==0.21.3
httpx==0.25.0
authlib==1.3.0
requests==2.31.0
gunicorn==21.2.0
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
mcp-python==0.1.0
playwright==1.48.0